test: refactor tests and fix failure
