+++ sputnikdao2/tests-ava/__tests__/proposals.ava.ts
@@ -13,8 +13,8 @@ import {
     initStaking,
     setStakingId,
     voteApprove,
-    initWorkspace,
     Proposal,
+    initWorkspace,
 } from './utils';
 
 const test = initWorkspace();
@@ -453,144 +453,6 @@ test(
         await voteApprove(root, dao, transferId);
         const { status } : Proposal = await dao.view('get_proposal', { id: transferId });
         t.is(status, 'InProgress');
-    },
-);
-
-initWorkspace({ noInit: true })(
-    'Proposal action types',
-    async (t) => {
-        const { alice, root, dao } = t.context.accounts;
-        const user1 = await root.createSubAccount('user1');
-        const user2 = await root.createSubAccount('user2');
-        const user3 = await root.createSubAccount('user3');
-        const period = new BN('1000000000')
-            .muln(60)
-            .muln(60)
-            .muln(24)
-            .muln(7)
-            .toString();
-        const policy = {
-            roles: [
-                {
-                    name: 'council',
-                    kind: {
-                        Group: [
-                            alice.accountId,
-                            user1.accountId,
-                            user2.accountId,
-                            user3.accountId,
-                        ],
-                    },
-                    permissions: ['*:*'],
-                    vote_policy: {},
-                },
-            ],
-            default_vote_policy: {
-                weight_kind: 'RoleWeight',
-                quorum: new BN('0').toString(),
-                threshold: [1, 2],
-            },
-            proposal_bond: toYocto('1'),
-            proposal_period: period,
-            bounty_bond: toYocto('1'),
-            bounty_forgiveness_period: period,
-        };
-
-        let config = { name: 'sputnik', purpose: 'testing', metadata: '' };
-
-        await root.call(dao, 'new', { config, policy });
-
-        let proposalId = await alice.call(
-            dao,
-            'add_proposal',
-            {
-                proposal: {
-                    description: 'rename the dao',
-                    kind: {
-                        ChangeConfig: {
-                            config,
-                        },
-                    },
-                },
-            },
-            { attachedDeposit: toYocto('1') },
-        );
-
-        // Remove proposal works
-        await alice.call(dao, 'act_proposal', {
-            id: proposalId,
-            action: 'RemoveProposal',
-        });
-        let err = await captureError(async () =>
-            dao.view('get_proposal', { id: proposalId }),
-        );
-        t.regex(err, /ERR_NO_PROPOSAL/);
-
-        err = await captureError(async () =>
-            alice.call(dao, 'act_proposal', {
-                id: proposalId,
-                action: 'VoteApprove',
-            }),
-        );
-        t.regex(err, /ERR_NO_PROPOSAL/);
-
-        proposalId = await alice.call(
-            dao,
-            'add_proposal',
-            {
-                proposal: {
-                    description: 'rename the dao',
-                    kind: {
-                        ChangeConfig: {
-                            config,
-                        },
-                    },
-                },
-            },
-            { attachedDeposit: toYocto('1') },
-        );
-
-        err = await captureError(async () =>
-            alice.call(dao, 'act_proposal', {
-                id: proposalId,
-                action: 'AddProposal',
-            }),
-        );
-        t.regex(err, /ERR_WRONG_ACTION/);
-
-        // Check if every vote counts
-        await user1.call(dao, 'act_proposal', {
-            id: proposalId,
-            action: 'VoteApprove',
-        });
-        await user2.call(dao, 'act_proposal', {
-            id: proposalId,
-            action: 'VoteReject',
-        });
-        await alice.call(dao, 'act_proposal', {
-            id: proposalId,
-            action: 'VoteRemove',
-        });
-        {
-            const { vote_counts, votes } = await dao.view('get_proposal', {
-                id: proposalId,
-            }) as any;
-            t.deepEqual(vote_counts.council, [1, 1, 1]);
-            t.deepEqual(votes, {
-                [alice.accountId]: 'Remove',
-                [user1.accountId]: 'Approve',
-                [user2.accountId]: 'Reject',
-            });
-        }
-
-        // Finalize proposal will panic if not exired or failed
-        err = await captureError(async () =>
-            alice.call(dao, 'act_proposal', {
-                id: proposalId,
-                action: 'Finalize',
-            }),
-        );
-        t.regex(err, /ERR_PROPOSAL_NOT_EXPIRED_OR_FAILED/);
     },
 );
